feat(ui-compo): add component to render common ui for all steps
The LoginForm component injects styling and renders children props along main sidebar

re #282
